fix(shadow): retire the superseded ESM task behind the frame fence too
Review follow-up: `ensureEsmShadowTaskState` disposed the previous task
synchronously exactly like the PCF hooks did, and it is reached through the
same `setShadowTaskCasterMeshes()` re-supply and frame scheduler, so a
mid-frame ESM caster replacement could still destroy buffers referenced by
the command buffer being recorded. Route it through `retireGpuResources`
as well, so every supported generator follows the same lifetime rule.

The regression test is parameterized over PCF and ESM (renamed to
shadow-task-retire.test.ts); the ESM builder is exported as @internal for it.
